+++ rust-runtime/aws-smithy-default-tls/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "aws-smithy-default-tls"
+authors = ["AWS Rust SDK Team <aws-sdk-rust@amazon.com>"]
+description = "Selects the default HTTPS stack for generated smithy clients based on platform support"
+version = "0.1.0"
+license = "Apache-2.0"
+edition = "2021"
+repository = "https://github.com/smithy-lang/smithy-rs"
+
+[dependencies]
+aws-smithy-runtime-api = { path = "../aws-smithy-runtime-api" }
+aws-smithy-http-client = { path = "../aws-smithy-http-client", optional = true }
+tracing = "0.1.40"
+
+[target.'cfg(unix)'.dependencies]
+aws-smithy-http-client = { path = "../aws-smithy-http-client", features = ["s2n-tls"] }
+
+[target.'cfg(windows)'.dependencies]
+aws-smithy-http-client = { path = "../aws-smithy-http-client", features = ["rustls-aws-lc"] }
+
+[package.metadata.docs.rs]
+all-features = true
+targets = ["x86_64-unknown-linux-gnu"]
+cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
+rustdoc-args = ["--cfg", "docsrs"]
+# End of docs.rs metadata
