+++ rust-runtime/aws-smithy-runtime/src/client/http.rs
@@ -2,9 +2,6 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-use aws_smithy_runtime_api::client::behavior_version::BehaviorVersion;
-#[cfg(feature = "default-https-client")]
-use aws_smithy_runtime_api::client::http::SharedHttpClient;
 
 /// Interceptor for connection poisoning.
 pub mod connection_poisoning;
@@ -55,41 +52,3 @@ pub mod hyper_014 {
 
 /// HTTP body and body-wrapper types
 pub mod body;
-
-// NOTE: We created default client options to evolve defaults over time (e.g. allow passing a different DNS resolver)
-/// Configuration options for the default HTTPS client
-#[derive(Debug, Clone)]
-pub(crate) struct DefaultClientOptions {
-    _behavior_version: BehaviorVersion,
-}
-
-impl Default for DefaultClientOptions {
-    fn default() -> Self {
-        DefaultClientOptions {
-            _behavior_version: BehaviorVersion::latest(),
-        }
-    }
-}
-
-impl DefaultClientOptions {
-    /// Set the behavior version to use
-    #[allow(unused)]
-    pub(crate) fn with_behavior_version(mut self, behavior_version: BehaviorVersion) -> Self {
-        self._behavior_version = behavior_version;
-        self
-    }
-}
-
-/// Creates an HTTPS client using the default TLS provider
-#[cfg(feature = "default-https-client")]
-pub(crate) fn default_https_client(_options: DefaultClientOptions) -> Option<SharedHttpClient> {
-    use aws_smithy_http_client::{tls, Builder};
-    tracing::trace!("creating a new default hyper 1.x client using rustls<aws-lc>");
-    Some(
-        Builder::new()
-            .tls_provider(tls::Provider::Rustls(
-                tls::rustls_provider::CryptoMode::AwsLc,
-            ))
-            .build_https(),
-    )
-}
